LayoutTests/fast/dom/navigator-property-gc-after-frame-detach-expected.txt

Tests that Navigator properties do not get GC'd before their Navigator object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS frameNavigator.geolocation.foo is 1
PASS frameNavigator.mimeTypes.foo is 1
PASS frameNavigator.plugins.foo is 1
PASS frameNavigator.serviceWorker.foo is 1

PASS frameNavigator.geolocation.foo is 1
PASS frameNavigator.mimeTypes.foo is 1
PASS frameNavigator.plugins.foo is 1
PASS frameNavigator.serviceWorker.foo is 1

PASS frameNavigator.geolocation.foo is 1
PASS frameNavigator.mimeTypes.foo is 1
PASS frameNavigator.plugins.foo is 1
PASS frameNavigator.serviceWorker.foo is 1
PASS successfullyParsed is true

TEST COMPLETE
